Move ProgressBar to separated component
