+++ src/components/progress-bar/index.ts
@@ -1,0 +1,7 @@
+// SPDX-FileCopyrightText: 2025 Istituto Nazionale di Fisica Nucleare
+//
+// SPDX-License-Identifier: EUPL-1.2
+
+export { useProgressBar } from "./context";
+export { ProgressBarProvider } from "./provider";
+export { ProgressBar } from "./progress-bar";
